requirements.txt
anyio==4.2.0
asgiref==3.7.2
azure-core==1.30.0
azure-storage-blob==12.19.0
Cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
defusedxml==0.7.1
Django==5.0.1
django-allauth==0.60.1
django-storages==1.14.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
isodate==0.6.1
msgpack==1.0.7
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
toolz==0.12.1
typing_extensions==4.9.0
tzdata==2023.4
ulid-py==1.1.0
urllib3==2.2.0
